Extract node context from member access
